// Music_Management_System/Music_Mangmt_System_UI/src/app/models/music.model.ts
export class Music {
    id: string;
    image: string = 'https://images.unsplash.com/photo-1619983081563-430f63602796?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Nnx8bXVzaWN8ZW58MHx8MHx8fDA%3D"';
    songName: string;
    singerName: string;
    releaseYear: number;
    movieName: string;
    songType: string;


}